==> __temp/0a Normalisasi huruf.ipynb <==
import pandas as pd
from unidecode import unidecode

import pandas as pd
import re, string, unicodedata
from unidecode import unidecode
data_replace  = pd.read_excel('data/Corpus_kata_replace_new.xlsx', sheet_name = 'kata_repalce')
load_spam = pd.read_excel('data/kata_spam.xlsx', sheet_name = 'Sheet1')
spam = list(set(load_spam['spam'].tolist()))
data_hapus  = pd.read_excel('data/Corpus_kata_replace_new.xlsx', sheet_name = 'kata_hapus')
#data_olah  = pd.read_excel('data/komentar_label_new.xlsx')
#data_new = data
diganti = data_replace['kata'].tolist()
ganti = data_replace['ganti'].tolist()
hapus = data_hapus['kata'].tolist()

fo = pd.read_excel('data/karakter.xlsx', sheet_name='Sheet1')
x = fo['karakter'].tolist() #
y = fo['replace'].tolist()
def gantiKarakter(str, x=x, y=y):
    for i in range(len(x)):
        if i == 0:
            n_word = str
        n_word = n_word.replace(x[i],y[i])
    return unidecode(n_word).lower()

def replace_sw(teks):
    teks_tokenize = teks.split()
    for j in range(len(teks_tokenize)):
        for k in range(len(diganti)):
            if teks_tokenize[j]==diganti[k]:
                teks_tokenize[j] = ganti[k]
        for k in range(len(hapus)):
            if teks_tokenize[j] == hapus[k]:
                teks_tokenize[j]= ''
    join = ' '.join(map(str,(teks_tokenize)))
    join = re.sub('[\s]+', ' ', join)
    return join

def removePunc(str):
    str = re.sub(r'[^\w]|_',' ',str)
    str = re.sub(r"\b\d+\b", " ", str)
    str = re.sub('[\s]+', ' ', str)
    return str
#x = removePunc(replace_sw(teks))


fo = pd.read_excel('data/karakter.xlsx', sheet_name='Sheet1')
x = fo['karakter'].tolist() #
y = fo['replace'].tolist()
def gantiKarakter(str, x=x, y=y):
    for i in range(len(x)):
        if i == 0:
            n_word = str
        n_word = n_word.replace(x[i],y[i])
    return unidecode(n_word).lower()

def normalAt(str):
    ok = gantiKarakter(str)
    n_w = []
    for i in range(len(ok)):
        if ok[i] == "@" and i !=0 and ok[i-1] !=" ":
            n_w.append(" @")
        else:
            n_w.append(ok[i])
    return "".join(n_w)

def cleaning(str):
    #remove non-ascii
    str = unicodedata.normalize('NFKD', str).encode('ascii', 'ignore').decode('utf-8', 'ignore')
    #remove URLs
    str = re.sub(r'(?i)\b((?:https?://|www\d{0,3}[.]|[a-z0-9.\-]+[.][a-z]{2,4}/)(?:[^\s()<>]+|\(([^\s()<>]+|(\([^\s()<>]+\)))*\))+(?:\(([^\s()<>]+|(\([^\s()<>]+\)))*\)|[^\s`!()\[\]{};:\'".,<>?«»“”‘’]))', '', str)
    str = str.lower()
    #Remove additional white spaces
    str = re.sub('[\s]+', ' ', str)
       
    return str
def getJtext(text):
    words = re.findall(r'[a-z0-9@.]+', text)
    return ' '.join(words)
def preprocessing1(str):
    text = cleaning(normalAt(str))
    text = getJtext(text)
    return text

for i in range(len(komentar)):
    x = (type(komentar[i]))
    if type(komentar[i]) is int:
    #if str(x) == "<class 'int'>":
        komentar[i] == "null"
        print(i)

komentar[2480]

komentar = km['komentar'].tolist()
n_komentar = []
for i in range(len(komentar)):
    if type(komentar[i]) is int:
        komentar[i] == "null"
    n_komentar.append(gantiKarakter(komentar[i]))
    #str = gantiKarakter(ll)
    #print(i)
    #n_komentar.append(gantiKarakter(i))


#dic = {"komentar":n_komentar}

#df = pd.DataFrame(dic,columns = ['komentar'])
#writer = pd.ExcelWriter('data/data_komentar2_normalisasi.xlsx', engine='xlsxwriter')
#df.to_excel(writer, sheet_name='Sheet1')
#writer.save()


data[128]

 from nltk.tokenize import TweetTokenizer
komentar[128]

data = pd.read_excel('data/Corpus_kata_replace_new.xlsx', sheet_name = "data_set")
komentar = data['komentar']
global str
for i in range(len(komentar)):
    if i >=130 and i <140:
        print(type(komentar[i]))
        #x = str(komentar[i])
        print(i)
        #print(cleaning(x))

data

#data = pd.read_excel('data/dataset_kotor3.xlsx', sheet_name = "Sheet1")
#komentar = data['komentar']
#kode = data['Column1']



n_komentar = []
for teks in range(len(komentar)):
    #komentar[teks]['sentiment']=komentar[teks].map_partitions(lambda df : df.apply(polar,axis=1))
    if type(komentar[teks]) is int:
        komentar = 'hapus'
    #print(teks)
    #proses = getJtext(komentar[teks])
    #proses = preprocessing1(komentar[teks])
    print(teks)
    proses = cleaning(komentar[teks])
    #proses = replace_sw(proses)
    n_komentar.append(proses)
#n2_komentar = preprocessing2(n_komentar)

komentar

kode_spam = []
#spam = x2
for i in n_komentar:
    k = tk.tokenize(i)
    dsp = "zero"
    for j in k:
        for s in spam:
            if s == j:
                dsp = "spam"
    if dsp == "spam":
        kode_spam.append(1)
    else:
        kode_spam.append("")
dic = {"kode":kode_spam}

#df = pd.DataFrame(dic,columns = ['kode'])
#writer = pd.ExcelWriter('data/kodelabel2.xlsx', engine='xlsxwriter')
#df.to_excel(writer, sheet_name='Sheet1')
#writer.save()
#df

ddd=0
for i in kode_spam:
    if i == 1:
        ddd+=1
print(ddd)


data = []
for i in range(len(n_komentar)):
    split = n_komentar[i].split()
    if len(split)>8 and  kode_spam[i] != 1:
        data.append(komentar[i])
dict_komentar  = {"komentar":data}
df = pd.DataFrame(dict_komentar,columns = ['komentar'])
writer = pd.ExcelWriter('data/komentar1_NO_SPAM.xlsx', engine='xlsxwriter')
df.to_excel(writer, sheet_name='Sheet1')
df.to_excel(writer, sheet_name='Sheet1')

# Close the Pandas Excel writer and output the Excel file.
writer.save()
len(data)

data = []
for i in range(len(n_komentar)):
    split = n_komentar[i].split()
    if len(split)>7 and  kode_spam[i] == 1:
        data.append(komentar[i])
dict_komentar  = {"komentar":data}
df = pd.DataFrame(dict_komentar,columns = ['komentar'])
writer = pd.ExcelWriter('data/komentar2_SPAM.xlsx', engine='xlsxwriter')
df.to_excel(writer, sheet_name='Sheet1')
df.to_excel(writer, sheet_name='Sheet1')

# Close the Pandas Excel writer and output the Excel file.
writer.save()
len(data)

df

s0 = """@ÁĶÚ MÁÚ ŃĞÁŚĨĤ ŤÁÚ @ŃĨĤ ßÚÁŤ ŤĔMĔŃ ŤĔMĔŃ ŶÁŃĞ Ŵ@ÁĴÁĤŃŶÁ ĶÚŚÁM ĎÁŃ ĴĔŔÁŴÁŤÁŃ@dr.dinda.fashionbeauty PÚŃŶÁ MÁŚĶĔŔ ŴÁĴÁĤ ŶÁŃĞ ÁMPÚĤ :) langsung aja Whatshapp 0895635365789
"""
def normalisasiT1(str):
    tk = TweetTokenizer()
    w = tk.tokenize(str)
    wo = []
    for i in w:
        if i == "," or i== "#" or i=="?" or i == "!" or i == ":":
            pass
        else:
            wo.append(i)
    return " ".join(wo)

normalisasiT1(s0)

def normalAt(str):
    ok = gantiKarakter(str)
    n_w = []
    for i in range(len(ok)):
        if ok[i] == "@" and i !=0 and ok[i-1] !=" ":
            n_w.append(" @")
        else:
            n_w.append(ok[i])
    return "".join(n_w)

def cleaning(str):
    #remove non-ascii
    str = unicodedata.normalize('NFKD', str).encode('ascii', 'ignore').decode('utf-8', 'ignore')
    #remove URLs
    str = re.sub(r'(?i)\b((?:https?://|www\d{0,3}[.]|[a-z0-9.\-]+[.][a-z]{2,4}/)(?:[^\s()<>]+|\(([^\s()<>]+|(\([^\s()<>]+\)))*\))+(?:\(([^\s()<>]+|(\([^\s()<>]+\)))*\)|[^\s`!()\[\]{};:\'".,<>?«»“”‘’]))', '', str)
    str = str.lower()
    #Remove additional white spaces
    str = re.sub('[\s]+', ' ', str)
       
    return str
def getJtext(text):
    words = re.findall(r'[a-z0-9@.]+', text)
    return ' '.join(words)
def preprocessing1(str):
    text = cleaning(normalAt(str))
    text = getJtext(text)
    return text

texsq ="Ғollowwww IG @BEAUTYSTORE_155  gaaaaa perluuuuυ repottt oleѕssss oleѕssss creammm,pυтιhhh pυnnnnn gaaaaa мeraтaaaaa,cυĸυpppoo мιnυnmmm  vιтalιneeeee dapatttt ĸannnn pυтιhhhh  мeraтaaaaa ѕelυrυhhhh тυвυнmhhhh .yυuuuu orderrrrr vιтalιneeeee nyaaaaa alwayssssѕ ѕтoĸkkkk тerвaтaѕsssss ..ιnғoooo  lengĸappppp wнaтѕнappppp : 087779738358@BEAUTYSTORE_155 💖💖"
import re, string, unicodedata
#print(normalAt(s0))
texs = cleaning(normalAt(texsq))
texs = getJtext(texs)
texs

import re
from collections import Counter

def words(text): return re.findall(r'\w+', text.lower())

WORDS = Counter(words(open('data/kata-dasar.original.txt').read()))

def P(word, N=sum(WORDS.values())): 
    # "Probability of `word`."
    return WORDS[word] / N

def correction(word): 
    # "Most probable spelling correction for word."
    return max(candidates(word), key=P)

def candidates(word): 
    # "Generate possible spelling corrections for word."
    return (known([word]) or known(edits1(word)) or known(edits2(word)) or [word])

def known(words): 
    # "The subset of `words` that appear in the dictionary of WORDS."
    return set(w for w in words if w in WORDS)

def edits1(word):
    # "All edits that are one edit away from `word`."
    letters    = 'abcdefghijklmnopqrstuvwxyz'
    splits     = [(word[:i], word[i:])    for i in range(len(word) + 1)] # [('', 'kemarin'), ('k', 'emarin'), ('ke', 'marin'), dst]
    deletes    = [L + R[1:]               for L, R in splits if R] # ['emarin', 'kmarin', 'kearin', dst]
    transposes = [L + R[1] + R[0] + R[2:] for L, R in splits if len(R)>1] # ['ekmarin', 'kmearin', 'keamrin', dst]
    replaces   = [L + c + R[1:]           for L, R in splits if R for c in letters] # ['aemarin', 'bemarin', 'cemarin', dst]
    inserts    = [L + c + R               for L, R in splits for c in letters] # ['akemarin', 'bkemarin', 'ckemarin', dst]
    return set(deletes + transposes + replaces + inserts)

def edits2(word): 
    # "All edits that are two edits away from `word`."
    return (e2 for e1 in edits1(word) for e2 in edits1(e1))

kata = 'kenalanku'
print('kata typo : ', kata)
print('koreksi : ', correction(kata))
