==> __temp/1b. seleksi data.ipynb <==
import re, string, unicodedata
import nltk
from nltk import word_tokenize, sent_tokenize
from nltk.corpus import stopwords

#preprocessing

def removeStopword(str):
    stop_words = set(stopwords.words('stopword_id_artikel_islam'))
    word_tokens = word_tokenize(str)
    filtered_sentence = [w for w in word_tokens if not w in stop_words]
    return ' '.join(filtered_sentence)

#remove sentence which contains only one word
def removeSentence(str): 
    word = str.split()
    wordCount = len(word)
    if(wordCount<=1):
        str = ''
    
    return str

def cleaning(str):
    #remove non-ascii
    str = unicodedata.normalize('NFKD', str).encode('ascii', 'ignore').decode('utf-8', 'ignore')
    #remove URLs
    str = re.sub(r'(?i)\b((?:https?://|www\d{0,3}[.]|[a-z0-9.\-]+[.][a-z]{2,4}/)(?:[^\s()<>]+|\(([^\s()<>]+|(\([^\s()<>]+\)))*\))+(?:\(([^\s()<>]+|(\([^\s()<>]+\)))*\)|[^\s`!()\[\]{};:\'".,<>?«»“”‘’]))', '', str)
    #remove punctuations
    str = re.sub(r'[^\w]|_',' ',str)
    #remove digit from string
    #str = re.sub("\S*\d\S*", "", str).strip()
    #remove digit or numbers
    #str = re.sub(r"\b\d+\b", " ", str)
    #to lowercase
    str = str.lower()
    #Remove additional white spaces
    str = re.sub('[\s]+', ' ', str)
       
    return str


def preprocessing(str):
    str = removeSentence(str)
    str = cleaning(str)
    #str = removeStopword(str)
    
    return str

cleaning("SsA 1 EWQEas1cs")

import pandas as pd
import xlsxwriter

fo = pd.read_excel('data/dataset_kotor2.xlsx', sheet_name='Sheet1') #read excel file
txt = fo['komentar'].tolist()
#workbook = xlsxwriter.Workbook('clean-data.xlsx')
#worksheet = workbook.add_worksheet()

data_baru = []
asli = []

for i in txt:
    data_baru.append(preprocessing(i))
data_baru1 = []
for i in range(len(data_baru)):
    if len(data_baru[i]) > 0:
        x = data_baru[i].split()
        if len(x)>3:
            data_baru1.append(data_baru[i])
            asli.append(txt[i])

dic_komentar = {"komentar":data_baru1, "asli":asli}
df = pd.DataFrame(dic_komentar,columns = ['komentar',"asli"])
writer = pd.ExcelWriter('data/kandidat_dataset.xlsx', engine='xlsxwriter')
df.to_excel(writer, sheet_name='Sheet1')
#df.to_excel(writer, sheet_name='Sheet1')

# Close the Pandas Excel writer and output the Excel file.
writer.save()
df

len(df)



fo = pd.read_excel('data/kandidat_dataset_label.xlsx', sheet_name='Sheet2')#read excel file
komentar = fo['komentar'].tolist()
label = fo['kode'].tolist()

k1= []
l1=[]
for i in range(len(komentar)):
    if len(komentar[i]) > 0:
        x = komentar[i].split()
        if len(x)>=7 or label[i]==1:
        #if len(x)>=5:
            k1.append(komentar[i])
            l1.append(label[i])
            
dic_komentar = {"komentar":k1, "label":l1}
df = pd.DataFrame(dic_komentar,columns = ['komentar',"label"])

writer = pd.ExcelWriter('data/kandidat_labeled.xlsx', engine='xlsxwriter')
df.to_excel(writer, sheet_name='Sheet1')
#df.to_excel(writer, sheet_name='Sheet1')

# Close the Pandas Excel writer and output the Excel file.
writer.save()
df

len(df)

data_replace  = pd.read_excel('data/Corpus_kata_replace_new.xlsx', sheet_name = 'kata_repalce')
data_hapus  = pd.read_excel('data/Corpus_kata_replace_new.xlsx', sheet_name = 'kata_hapus')
#data_olah  = pd.read_excel('data/komentar_label_new.xlsx')
#data_new = data
diganti = data_replace['kata'].tolist()
ganti = data_replace['ganti'].tolist()
hapus = data_hapus['kata'].tolist()

sp = []
for i in range(len(k1)):
    x = k1[i].split()
    for j in x:
        sp.append(j) 
setl = sorted(list(set(sp)))
#dic = {"komentar":sp}
#df = pd.DataFrame(dic,columns = ['komentar'])
len(setl)

#vocab = []
helper = []
for i in setl:
    for j in diganti:
        if j == i:
            helper.append(j)
for i in setl:
    for j in helper:
        if j ==

voxB  = list(set(setl) - set(helper))
print(len(setl))
print(len(voxB))

for i in diganti:
    for j in voxB:
        if i == j:
            print(i)
dic = {"kata":voxB}
df = pd.DataFrame(dic,columns = ['kata'])
writer = pd.ExcelWriter('data/list_vocab_backup.xlsx', engine='xlsxwriter')
df.to_excel(writer, sheet_name='Sheet1')
#df.to_excel(writer, sheet_name='Sheet1')

# Close the Pandas Excel writer and output the Excel file.
writer.save()
df
