==> __temp/5a Implementasi - TF-IDF.ipynb <==
import pickle
import pandas as pd
import time
import numpy as np
import re
from sklearn.feature_extraction.text import TfidfVectorizer
from Sastrawi.StopWordRemover.StopWordRemoverFactory import StopWordRemoverFactory
from Sastrawi.Stemmer.StemmerFactory import StemmerFactory

#Fungsi cleaning
def cleaning2(text):
    text = re.sub(r'\b(?:(?:https?|ftp)://)?\w[\w-]*(?:\.[\w-]+)+\S*(?<![.,])', ' ', text.lower())
    words = re.findall(r'[a-z0-9]+', text)
    return ' '.join(words)

    #fungsi hapus
def hapus(n_komentar):
    clean_komentar = []
    clean_komentar_hapus = []
    for i in range(len(n_komentar)):
        x = str(n_komentar[i])
        splt = x.split(" ")
        for j in range(len(splt)):
            for k in range(len(diganti)):
                if splt[j] == diganti[k]:
                    splt[j]= ganti[k]
        join = ' '.join(map(str,(splt)))
        clean_komentar.append(join)
    
    for i in range(len(clean_komentar)):
        x = str(clean_komentar[i])
        splt = x.split(" ")
        for j in range(len(splt)):
            for k in range(len(hapus)):
                if splt[j] == hapus[k]:
                    splt[j]= ''
        join = ' '.join(map(str,(splt)))
        clean_komentar_hapus.append(join)

def normalisasi(n_komentar):
    clean_komentar = []
    clean_komentar_hapus = []
    for i in range(len(n_komentar)):
        x = str(n_komentar[i]) 
        splt = x.split(" ")
        for j in range(len(splt)):
            for k in range(len(diganti)):
                if splt[j] == diganti[k]:
                    splt[j]= ganti[k]
        join = ' '.join(map(str,(splt)))
        clean_komentar.append(join)
    
    for i in range(len(clean_komentar)):
        x = str(clean_komentar[i])
        splt = x.split(" ")
        for j in range(len(splt)):
            for k in range(len(hapuss)):
                if splt[j] == hapuss[k]:
                    splt[j]= ''
        join = ' '.join(map(str,(splt)))
        clean_komentar_hapus.append(join)
    return str(clean_komentar_hapus)
    
# sastrawi untuk proses stemming
factory = StemmerFactory()
factoryStop = StopWordRemoverFactory()
stemmer = factory.create_stemmer()
stopword = factoryStop.create_stop_word_remover()

#load dataset untk normalisasi
data_replace  = pd.read_excel('data/Corpus_kata_replace_new.xlsx', sheet_name = 'kata_repalce')
data_hapus  = pd.read_excel('data/Corpus_kata_replace_new.xlsx', sheet_name = 'kata_hapus')
diganti = data_replace['kata'].tolist()
ganti = data_replace['ganti'].tolist()
hapuss = data_hapus['kata'].tolist()


# load the features from disk
x = open('model/feature.txt')
x = x.read()
koso_kata = x.split()
tfidf = TfidfVectorizer(vocabulary=koso_kata)

# load the model from disk
comNB = pickle.load(open('model\comNB.sav', 'rb'))
cSVM = pickle.load(open('model\cSVM.sav', 'rb'))


#input_komentar = "Cek IG kami Kak, solusiii naiiikkk tinggiiii sampe 175cm, bahkan sampe umuur 33th masih bisa looh, Buruaaaaannn. '"
input_komentar = input("Komentar: ")
#komentar  = [cleaning2(input_komentar)]
komentar  = cleaning2(input_komentar)
print("==============================================================================================================================")
print(str(komentar))
print("==============================================================================================================================")
komentar = normalisasi(komentar)
komentar = stemmer.stem(komentar)
komentar = stopword.remove(komentar)

t_komentar = []
t_komentar.append(komentar)
t_tfidf_matrix = tfidf.fit_transform(t_komentar)

data_X = t_tfidf_matrix.toarray()

x_cek = data_X[len(t_komentar)-1:len(t_komentar)]

# prediksi data asli
Y_SVM = cSVM.predict(x_cek)
Y_NB = comNB.predict(x_cek)

# print prediksi
print("Prediksi SVM (TF-IDF)           : ", Y_SVM)
print("Prediksi C Naive Bayes (TF-IDF) : ", Y_NB)
print(t_tfidf_matrix.shape)
